requirements.txt
gradio>=4.26.0
requests>=2.31.0
Pillow>=10.1.0
python-dotenv>=1.0.0 